Create List-PCS-with-Specific-USB-Device.ps1
This script uses a selected Text file to use CIM to query Multiple PCs in the effort to find a specific usb device that you name. SCCM is not always the most reliable, so I create this quick dirty report to tell me where stuff is. Especially when it "magically" moves around in an environment. You can pair it with other scripts of mine to "get the details". And figure out whom is the one buying lunch today, lol . I can also create a separate script to help you list the names of a specfic computer name. Just feel free to email me if you want want differrent mods for your environment.  Or fill out my contact form on https://jlaforce.me and I'll get back to ya.